+++ src/store/shared/index.js
@@ -1,0 +1,34 @@
+
+export default{
+  state: {
+    loading: false,
+    error: null
+  },
+  mutations: {
+    
+    setLoading(state, payload) {
+      state.loading = payload;
+    },
+    setError(state, payload) {
+      state.error = payload;
+    },
+    clearError(state) {
+      state.error = null;
+    }
+      },
+  actions: {
+ 
+    clearError({ commit }) {
+      commit("clearError");
+    }
+  },
+  getters: {
+   
+    error(state) {
+      return state.error;
+    },
+    loading(state) {
+      return state.loading;
+  }
+  }
+}
